Deduplication of the inner SaveOnAccessCsrfToken classes
